Added new base EntityTopicOperatorSpec class Inherited TopicOperatorSpec from above for backward compatibility Refactored UserOperatorSpec into EntityOperatorSpec and changed timeouts in seconds
